requirements.txt
Flask>=2.0.2
Flask-Cors==3.0.10
requests==2.27.1
python-dotenv==1.0.0